+++ .eslintrc
@@ -1,0 +1,31 @@
+env:
+  node: true
+  browser: true
+
+ecmaFeatures:
+  jsx: true
+
+plugins:
+  - react
+
+rules:
+  camelcase: 2
+  eol-last: 2
+  quotes: [2, "single", "avoid-escape"]
+  no-mixed-spaces-and-tabs: 2
+  no-trailing-spaces: 2
+  no-underscore-dangle: 0
+  no-unused-vars: [1, { "args": "none" }]
+  vars-on-top: 1
+  react/jsx-boolean-value: 1
+  react/jsx-quotes: [2, "double", "avoid-escape"]
+  react/jsx-no-undef: 1
+  react/jsx-sort-prop-types: 1
+  react/jsx-uses-vars: 1
+  react/no-did-mount-set-state: 1
+  react/no-did-update-set-state: 1
+  react/no-multi-comp: 1
+  react/no-unknown-property: 1
+  react/prop-types: 1
+  react/self-closing-comp: 1
+  react/wrap-multilines: 1
