+++ .travis.yml
@@ -1,8 +1,9 @@
 language: node_js
 node_js:
   - "0.10"
-  
-sudo: false  
+  - "0.12"
+
+sudo: false
 
 script:
   - "npm run-script ci"
