+++ __tests__/test.js
@@ -1,14 +1,14 @@
 // __tests__/test.js
 
-jest.dontMock('../src/react_components/facebook.jsx');
+jest.dontMock('../src/react_components/facebook');
 
 describe('test', function() {
   it('the widget actually renders', function() {
     var React = require('react/addons');
     var TestUtils = React.addons.TestUtils;
-    var Facebook = require('../src/react_components/facebook.jsx');
-    var title = "mullet";
-    var subtitle = "stuff";
+    var Facebook = require('../src/react_components/facebook');
+    var title = 'mullet';
+    var subtitle = 'stuff';
 
     // Render a checkbox with label in the document
     var facebook = TestUtils.renderIntoDocument(
