+++ package.json
@@ -30,26 +30,35 @@
   },
   "homepage": "http://www.mullet.io",
   "dependencies": {
+    "babel": "^5.6.23",
+    "babel-jest": "^5.3.0",
+    "babelify": "^6.1.0",
     "body-parser": "1.0.2",
     "bower": "^1.3.12",
-    "browserify": "^8.1.1",
+    "browserify": "^10.2.6",
     "grunt": "^0.4.5",
-    "grunt-browserify": "^3.3.0",
+    "grunt-browserify": "^3.8.0",
     "grunt-cli": "^0.1.13",
-    "grunt-contrib-jshint-jsx": "^0.10.2",
+    "grunt-contrib-eslint": "0.0.5",
     "grunt-contrib-watch": "^0.6.1",
-    "grunt-react": "^0.9.0",
-    "react": "^0.13.1",
-    "hapi": "^8.4.0",
-    "react-tools": "^0.13.1"
+    "hapi": "^8.8.0",
+    "react": "^0.13.3"
   },
   "devDependencies": {
-    "jest-cli": "^0.4.0"
+    "eslint": "^1.0.0-rc-1",
+    "eslint-plugin-react": "^2.2.0",
+    "jest-cli": "^0.4.15"
   },
   "jest": {
-    "scriptPreprocessor": "./preprocessor.js",
-    "unmockedModulePathPatterns": [
-      "./node_modules/react"
+    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
+    "testFileExtensions": [
+      "es6",
+      "js"
+    ],
+    "moduleFileExtensions": [
+      "js",
+      "json",
+      "es6"
     ]
   }
 }
